# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MaintSys API",
        "description": "API de gerenciamento de manutenção industrial - MaintSys",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/status-alerts": {
            "get": {
                "tags": [
                    "Alertas de Status"
                ],
                "summary": "Lista todos os alertas de status",
                "operationId": "a8d25cb4b9b686ecc54e69a1d3b109b2",
                "responses": {
                    "200": {
                        "description": "Lista de alertas de status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "machine_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "previous_status": {
                                                "type": "string",
                                                "example": "active"
                                            },
                                            "new_status": {
                                                "type": "string",
                                                "example": "maintenance"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Máquina entrou em manutenção"
                                            },
                                            "triggered_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-10T09:00:00Z"
                                            },
                                            "is_read": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "machine": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alertas de Status"
                ],
                "summary": "Cria um novo alerta de status",
                "operationId": "e409627d0f8abc669537e770636aa24b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "machine_id",
                                    "previous_status",
                                    "new_status",
                                    "triggered_at"
                                ],
                                "properties": {
                                    "machine_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "previous_status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "new_status": {
                                        "type": "string",
                                        "example": "inactive"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Máquina desligada por falta de energia"
                                    },
                                    "triggered_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-10T09:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alerta criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "machine_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "previous_status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "new_status": {
                                            "type": "string",
                                            "example": "inactive"
                                        },
                                        "is_read": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "machine": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/status-alerts/{id}": {
            "get": {
                "tags": [
                    "Alertas de Status"
                ],
                "summary": "Retorna um alerta de status específico",
                "operationId": "7662693714a35846b8dfacfc0c7694f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do alerta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do alerta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "machine_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "previous_status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "new_status": {
                                            "type": "string",
                                            "example": "maintenance"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Máquina entrou em manutenção"
                                        },
                                        "triggered_at": {
                                            "type": "string",
                                            "example": "2024-01-10T09:00:00Z"
                                        },
                                        "is_read": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "machine": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Alerta não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alertas de Status"
                ],
                "summary": "Atualiza um alerta de status",
                "operationId": "441cc02eb4db27b8b5142082537ca8df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do alerta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Alerta revisado pela equipe"
                                    },
                                    "is_read": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alerta atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Alerta revisado pela equipe"
                                        },
                                        "is_read": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Alerta não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alertas de Status"
                ],
                "summary": "Remove um alerta de status",
                "operationId": "91053d2dd5a4f4c3be2f9b5ada24850d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do alerta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alerta removido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Alerta não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}